CG-6243: Review, fix issue and refactor (#43)
# CG-6243: Review, fix issue and refactor

* Add debug config file
* move refreshAccessToken function to oauth.ts
* refreshAccessToken changes:
* * change param from query string to request body -> this fix the issue
that cause bad request to token endpoint
* * the token contains provider to detect proper providerConfig
* * remove `clientSecret` from `okta` and `optimizely_cms` because it
causes to bad request. Those two providers don't need clientSecret
* nextauth.ts changes:
* * move accessTokenParse to EpiserverOIDCProvider
* * `accessTokenExpires` is calculated as `account.expires_at * 1000`
because expires_in is not existing in the account.
* Remove redundant code

## Issue:
* The jwt callback or refreshAccessToken is triggered twice and that
last jwt callback still holds the old token value. However the session
is up to date so we can skip this issue at the moment.
https://github.com/nextauthjs/next-auth/discussions/1455